refactor: create @pnpm/build-modules
