+++ packages/build-modules/README.md
@@ -1,0 +1,17 @@
+# @pnpm/build-modules
+
+> Build packages in node_modules
+
+<!--@shields('npm')-->
+[![npm version](https://img.shields.io/npm/v/@pnpm/build-modules.svg)](https://www.npmjs.com/package/@pnpm/build-modules)
+<!--/@-->
+
+## Installation
+
+```sh
+npm i -S @pnpm/build-modules
+```
+
+## License
+
+[MIT](./LICENSE) © [Zoltan Kochan](https://www.kochan.io/)
