+++ packages/build-modules/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "@pnpm/build-modules",
+  "version": "0.0.0",
+  "description": "Build packages in node_modules",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": ">=8"
+  },
+  "scripts": {
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
+    "tsc": "rimraf lib && tsc",
+    "test": "npm run tsc && npm run lint && mos t",
+    "md": "mos",
+    "prepublishOnly": "npm run tsc"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/build-modules",
+  "keywords": [
+    "pnpm",
+    "resolver",
+    "npm"
+  ],
+  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://pnpm.js.org",
+  "dependencies": {
+    "@pnpm/constants": "1.0.0",
+    "@pnpm/core-loggers": "3.0.0",
+    "@pnpm/lifecycle": "5.0.0",
+    "@pnpm/read-package-json": "2.0.1",
+    "@pnpm/store-controller-types": "3.0.0",
+    "@types/node": "*",
+    "@types/ramda": "0.26.6",
+    "graph-sequencer": "2.0.0",
+    "run-groups": "2.0.0"
+  },
+  "devDependencies": {
+    "@pnpm/build-modules": "link:",
+    "@pnpm/logger": "2.1.0",
+    "@pnpm/tslint-config": "0.0.0",
+    "mos": "2.0.0-alpha.3",
+    "mos-plugin-readme": "1.0.4",
+    "rimraf": "2.6.3",
+    "ts-node": "8.0.3",
+    "tslint": "5.15.0",
+    "typescript": "3.4.1"
+  },
+  "mos": {
+    "plugins": [
+      "readme"
+    ],
+    "installation": {
+      "useShortAlias": true
+    }
+  }
+}
