+++ packages/build-modules/typings/index.d.ts
@@ -1,0 +1,9 @@
+declare module 'read-package-json' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'graph-sequencer' {
+  const anything: any;
+  export = anything;
+}
