+++ packages/supi/src/install/resolvePeers.ts
@@ -42,7 +42,7 @@ export interface DependenciesGraphNode {
   prod: boolean,
   dev: boolean,
   optional: boolean,
-  id: string,
+  packageId: string,
   installable: boolean,
   additionalInfo: {
     deprecated?: string,
@@ -243,7 +243,6 @@ function resolvePeersOfNode (
       fetchingRawManifest: node.resolvedPackage.fetchingRawManifest,
       hasBin: node.resolvedPackage.hasBin,
       hasBundledDependencies: node.resolvedPackage.hasBundledDependencies,
-      id: node.resolvedPackage.id,
       independent,
       installable: node.installable,
       isBuilt: !!node.resolvedPackage.engineCache,
@@ -252,6 +251,7 @@ function resolvePeersOfNode (
       name: node.resolvedPackage.name,
       optional: node.resolvedPackage.optional,
       optionalDependencies: node.resolvedPackage.optionalDependencies,
+      packageId: node.resolvedPackage.id,
       peripheralLocation,
       prepare: node.resolvedPackage.prepare,
       prod: node.resolvedPackage.prod,
